修改了PID文件 和OFFSET函数中的错误

+++ rsi/src/cg_src/offset_calculate.h
@@ -8,8 +8,8 @@
 #ifndef CG_SRC_OFFSET_CALCULATE_H_
 #define CG_SRC_OFFSET_CALCULATE_H_
 
-float rasAngleCalculate(uint8_t offset);
-float rasY_offsetCalculate(uint8_t openmv_x_offset);
-float rasX_offsetCalculate(uint8_t openmv_y_offset);
+float rasAngleCalculate(float offset);
+float rasY_offsetCalculate(float openmv_x_offset);
+float rasX_offsetCalculate(float openmv_y_offset);
 
 #endif /* CG_SRC_OFFSET_CALCULATE_H_ */
